requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiohttp==3.9.1
APScheduler==3.10.4
pydantic==2.5.3
